requirements.txt
numpy==1.24.3
matplotlib==3.7.1
scipy==1.10.1
Pillow>=9.5.0